Header animation bug fix
