+++ resources/assets/js/app.js
@@ -1,1 +1,2 @@
 require('./bootstrap');
+require('./scripts')
