+++ resources/views/site/header.blade.php
@@ -1,7 +1,12 @@
 <div id="main-header">
-	<a href="/">
+	<a href="/" class="non-sticky-area">
 		<span class="logo-prefix">Floppy</span>
-		<img class="avatar" src="/img/floppy-disk.svg" alt="FloppyDev" style="float: left;">
+		<img class="avatar" src="/img/floppy-disk.svg" alt="FloppyDev">
+		<span class="logo-suffix">Dev</span>
+	</a>
+	<a class="sticky-area" href="/">
+		<img class="avatar" src="/img/floppy-disk.svg" alt="FloppyDev">
+		<span class="logo-prefix">Floppy</span>
 		<span class="logo-suffix">Dev</span>
 	</a>
 </div>
